+++ test/adapter/json_api/has_many_test.rb
@@ -129,7 +129,7 @@ module ActiveModelSerializers
           assert_equal({
             data: {
               id: '1',
-              type: 'virtual_values',
+              type: 'virtual-values',
               relationships: {
                 maker: { data: { id: 1 } },
                 reviews: { data: [{ id: 1 }, { id: 2 }] }