+++ test/adapter/json_api/pagination_links_test.rb
@@ -25,7 +25,6 @@ module ActiveModelSerializers
           context = Minitest::Mock.new
           context.expect(:request_url, original_url)
           context.expect(:query_parameters, query_parameters)
-          context.expect(:key_transform, nil)
           @options = {}
           @options[:serialization_context] = context
         end