+++ test/adapter/json_api/transform_test.rb
@@ -36,13 +36,13 @@ module ActiveModelSerializers
           belongs_to :author
         end
 
-        def mock_request(key_transform = nil)
+        def mock_request(transform = nil)
           context = Minitest::Mock.new
           context.expect(:request_url, URI)
           context.expect(:query_parameters, {})
-          context.expect(:key_transform, key_transform)
           context.expect(:url_helpers, Rails.application.routes.url_helpers)
           @options = {}
+          @options[:key_transform] = transform if transform
           @options[:serialization_context] = context
         end
 
@@ -64,7 +64,7 @@ module ActiveModelSerializers
           @comment2.post = @post
         end
 
-        def test_success_document_key_transform_default
+        def test_success_document_transform_default
           mock_request
           serializer = PostSerializer.new(@post)
           adapter = ActiveModelSerializers::Adapter::JsonApi.new(serializer)
@@ -98,7 +98,7 @@ module ActiveModelSerializers
           }, result)
         end
 
-        def test_success_document_key_transform_global_config
+        def test_success_document_transform_global_config
           mock_request
           result = with_config(key_transform: :camel_lower) do
             serializer = PostSerializer.new(@post)
@@ -134,7 +134,7 @@ module ActiveModelSerializers
           }, result)
         end
 
-        def test_success_doc_key_transform_serialization_ctx_overrides_global
+        def test_success_doc_transform_serialization_ctx_overrides_global
           mock_request(:camel)
           result = with_config(key_transform: :camel_lower) do
             serializer = PostSerializer.new(@post)
@@ -144,7 +144,7 @@ module ActiveModelSerializers
           assert_equal({
             Data: {
               Id: '1337',
-              Type: 'posts',
+              Type: 'Posts',
               Attributes: {
                 Title: 'Title 1',
                 Body: 'Body 1',
@@ -152,12 +152,12 @@ module ActiveModelSerializers
               },
               Relationships: {
                 Author: {
-                  Data: { Id: '1', Type: 'authors' }
+                  Data: { Id: '1', Type: 'Authors' }
                 },
                 Comments: {
                   Data: [
-                    { Id: '7', Type: 'comments' },
-                    { Id: '12', Type: 'comments' }
+                    { Id: '7', Type: 'Comments' },
+                    { Id: '12', Type: 'Comments' }
                 ] }
               },
               Links: {
@@ -170,8 +170,8 @@ module ActiveModelSerializers
           }, result)
         end
 
-        def test_success_document_key_transform_dashed
-          mock_request(:dashed)
+        def test_success_document_transform_dash
+          mock_request(:dash)
           serializer = PostSerializer.new(@post)
           adapter = ActiveModelSerializers::Adapter::JsonApi.new(serializer)
           result = adapter.serializable_hash(@options)
@@ -204,7 +204,7 @@ module ActiveModelSerializers
           }, result)
         end
 
-        def test_success_document_key_transform_unaltered
+        def test_success_document_transform_unaltered
           mock_request(:unaltered)
           serializer = PostSerializer.new(@post)
           adapter = ActiveModelSerializers::Adapter::JsonApi.new(serializer)
@@ -238,7 +238,7 @@ module ActiveModelSerializers
           }, result)
         end
 
-        def test_success_document_key_transform_undefined
+        def test_success_document_transform_undefined
           mock_request(:zoot)
           serializer = PostSerializer.new(@post)
           adapter = ActiveModelSerializers::Adapter::JsonApi.new(serializer)
@@ -247,7 +247,7 @@ module ActiveModelSerializers
           end
         end
 
-        def test_success_document_key_transform_camel
+        def test_success_document_transform_camel
           mock_request(:camel)
           serializer = PostSerializer.new(@post)
           adapter = ActiveModelSerializers::Adapter::JsonApi.new(serializer)
@@ -255,7 +255,7 @@ module ActiveModelSerializers
           assert_equal({
             Data: {
               Id: '1337',
-              Type: 'posts',
+              Type: 'Posts',
               Attributes: {
                 Title: 'Title 1',
                 Body: 'Body 1',
@@ -263,12 +263,12 @@ module ActiveModelSerializers
               },
               Relationships: {
                 Author: {
-                  Data: { Id: '1', Type: 'authors' }
+                  Data: { Id: '1', Type: 'Authors' }
                 },
                 Comments: {
                   Data: [
-                    { Id: '7', Type: 'comments' },
-                    { Id: '12', Type: 'comments' }
+                    { Id: '7', Type: 'Comments' },
+                    { Id: '12', Type: 'Comments' }
                 ] }
               },
               Links: {
@@ -281,7 +281,7 @@ module ActiveModelSerializers
           }, result)
         end
 
-        def test_success_document_key_transform_camel_lower
+        def test_success_document_transform_camel_lower
           mock_request(:camel_lower)
           serializer = PostSerializer.new(@post)
           adapter = ActiveModelSerializers::Adapter::JsonApi.new(serializer)
@@ -315,7 +315,7 @@ module ActiveModelSerializers
           }, result)
         end
 
-        def test_error_document_key_transform_default
+        def test_error_document_transform_default
           mock_request
           resource = ModelWithErrors.new
           resource.errors.add(:published_at, 'must be in the future')
@@ -327,7 +327,7 @@ module ActiveModelSerializers
             { :errors =>
               [
                 {
-                  :source => { :pointer => '/data/attributes/published_at' },
+                  :source => { :pointer => '/data/attributes/published-at' },
                   :detail => 'must be in the future' },
                 {
                   :source => { :pointer => '/data/attributes/title' },
@@ -338,7 +338,7 @@ module ActiveModelSerializers
           assert_equal expected_errors_object, result
         end
 
-        def test_error_document_key_transform_global_config
+        def test_error_document_transform_global_config
           mock_request
           result = with_config(key_transform: :camel) do
             resource = ModelWithErrors.new
@@ -352,11 +352,11 @@ module ActiveModelSerializers
             { :Errors =>
               [
                 {
-                  :Source => { :Pointer => '/data/attributes/published_at' },
+                  :Source => { :Pointer => '/data/attributes/PublishedAt' },
                   :Detail => 'must be in the future'
                 },
                 {
-                  :Source => { :Pointer => '/data/attributes/title' },
+                  :Source => { :Pointer => '/data/attributes/Title' },
                   :Detail => 'must be longer'
                 }
               ]
@@ -364,7 +364,7 @@ module ActiveModelSerializers
           assert_equal expected_errors_object, result
         end
 
-        def test_error_document_key_transform_serialization_ctx_overrides_global
+        def test_error_document_transform_serialization_ctx_overrides_global
           mock_request(:camel)
           result = with_config(key_transform: :camel_lower) do
             resource = ModelWithErrors.new
@@ -378,11 +378,11 @@ module ActiveModelSerializers
             { :Errors =>
               [
                 {
-                  :Source => { :Pointer => '/data/attributes/published_at' },
+                  :Source => { :Pointer => '/data/attributes/PublishedAt' },
                   :Detail => 'must be in the future'
                 },
                 {
-                  :Source => { :Pointer => '/data/attributes/title' },
+                  :Source => { :Pointer => '/data/attributes/Title' },
                   :Detail => 'must be longer'
                 }
               ]
@@ -390,8 +390,8 @@ module ActiveModelSerializers
           assert_equal expected_errors_object, result
         end
 
-        def test_error_document_key_transform_dashed
-          mock_request(:dashed)
+        def test_error_document_transform_dash
+          mock_request(:dash)
 
           resource = ModelWithErrors.new
           resource.errors.add(:published_at, 'must be in the future')
@@ -405,7 +405,7 @@ module ActiveModelSerializers
             { :errors =>
               [
                 {
-                  :source => { :pointer => '/data/attributes/published_at' },
+                  :source => { :pointer => '/data/attributes/published-at' },
                   :detail => 'must be in the future'
                 },
                 {
@@ -417,7 +417,7 @@ module ActiveModelSerializers
           assert_equal expected_errors_object, result
         end
 
-        def test_error_document_key_transform_unaltered
+        def test_error_document_transform_unaltered
           mock_request(:unaltered)
 
           resource = ModelWithErrors.new
@@ -438,7 +438,7 @@ module ActiveModelSerializers
           assert_equal expected_errors_object, result
         end
 
-        def test_error_document_key_transform_undefined
+        def test_error_document_transform_undefined
           mock_request(:krazy)
 
           resource = ModelWithErrors.new
@@ -453,7 +453,7 @@ module ActiveModelSerializers
           end
         end
 
-        def test_error_document_key_transform_camel
+        def test_error_document_transform_camel
           mock_request(:camel)
 
           resource = ModelWithErrors.new
@@ -467,14 +467,14 @@ module ActiveModelSerializers
           expected_errors_object =
             { :Errors =>
               [
-                { :Source => { :Pointer => '/data/attributes/published_at' }, :Detail => 'must be in the future' },
-                { :Source => { :Pointer => '/data/attributes/title' }, :Detail => 'must be longer' }
+                { :Source => { :Pointer => '/data/attributes/PublishedAt' }, :Detail => 'must be in the future' },
+                { :Source => { :Pointer => '/data/attributes/Title' }, :Detail => 'must be longer' }
               ]
           }
           assert_equal expected_errors_object, result
         end
 
-        def test_error_document_key_transform_camel_lower
+        def test_error_document_transform_camel_lower
           mock_request(:camel_lower)
 
           resource = ModelWithErrors.new
@@ -488,7 +488,7 @@ module ActiveModelSerializers
           expected_errors_object =
             { :errors =>
               [
-                { :source => { :pointer => '/data/attributes/published_at' }, :detail => 'must be in the future' },
+                { :source => { :pointer => '/data/attributes/publishedAt' }, :detail => 'must be in the future' },
                 { :source => { :pointer => '/data/attributes/title' }, :detail => 'must be longer' }
               ]
           }